Give S3Bucket the ability to send PUT requests
Copy most of the code from the PUTObject class. I'm still trying to find
an intuitive and easy-to-use interface for the Amazon S3 REST API, and
this feels like a step in the right direction. A separate PUTObject
class seems unnecessary and a bit unwieldy to me at the moment.