Use multithreaded uploading with a progress bar
This is a major refactoring. Files are now uploaded in parallel in their
own threads and a progress bar shows how many bytes have been uploaded
(as well as how many are left).

NOTE: The multithreaded code introduced in this commit is VERY naive: it
simply creates a new thread for each image it finds in the specified
directory. Needless to say this may not behave well in directories that
contain a large number of files.

POSSIBLE BUG: The program once quit with an error about a detected
deadlock. I may need to fix this in a later commit.